+++ static_report/src/components/SingleReport/SRAssertionsSummaryLabel.tsx
@@ -1,0 +1,39 @@
+import { Flex, Text, Icon } from '@chakra-ui/react';
+import { FiCheck, FiX } from 'react-icons/fi';
+
+import { type AssertionTest } from '../../sdlc/single-report-schema';
+
+export function SRAssertionsSummaryLabel({
+  assertions,
+}: {
+  assertions: AssertionTest[];
+}) {
+  const total = assertions.length;
+  const failed = assertions.reduce((acc, test) => {
+    if (test.status === 'failed') {
+      acc++;
+    }
+    return acc;
+  }, 0);
+  const isPassed = failed === 0;
+
+  return (
+    <Flex gap={2} alignItems="center">
+      <Flex
+        alignItems="center"
+        borderRadius="md"
+        bgColor={isPassed ? '#DEFFEB' : '#FFE8F0'}
+        color={isPassed ? '#1F7600' : '#F60059'}
+        py={0.5}
+        px={1.5}
+      >
+        <Icon as={isPassed ? FiCheck : FiX} boxSize={4} />
+        <Text as="span">{isPassed ? 'All' : failed}</Text>
+      </Flex>
+      <Text as="span" color="gray.500">
+        of
+      </Text>
+      <Text as="span">{total}</Text>
+    </Flex>
+  );
+}
