+++ static_report/src/components/SingleReport/SRColumnLabel.tsx
@@ -1,0 +1,23 @@
+import { Flex, Center, Text, Icon } from '@chakra-ui/react';
+import { FiGrid } from 'react-icons/fi';
+
+export function SRColumnLabel({
+  name,
+  visibleIcon = true,
+  icon = FiGrid,
+}: {
+  visibleIcon?: boolean;
+  name: string;
+  icon?: any;
+}) {
+  return (
+    <Flex borderRadius="md" bgColor="gray.100" py={0.5} px={1}>
+      <Center>
+        {visibleIcon && <Icon as={icon} color="piperider.500" mr={1} />}
+        <Text as="span" fontSize="sm" color="gray.600">
+          {name}
+        </Text>
+      </Center>
+    </Flex>
+  );
+}
